requirements.txt
reflex>=0.4.0
cx_Oracle>=8.3.0
pydantic>=2.0.0
bcrypt>=4.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
datetime
uuid
pathlib
